+++ example/main.dart
@@ -2,7 +2,8 @@ import 'package:chem_nor/chem_nor.dart';
 
 void main() async {
   print(ChemNOR.availableModels);
-  final finder = ChemNOR(genAiApiKey: "key", model: GeminiModel.gemini3_0Flash);
+  final finder =
+      ChemNOR(genAiApiKey: "key", model: GeminiModel.gemini3_0Flash.apiName);
   dynamic properties = await finder.getCompoundProperties(248);
   print(properties);
   dynamic list = await finder.getSubstructureCids('CC');