+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.5.2
+## 0.5.7
 
-- updating the models to support the gemini 3 models.
-- debugging hints added.
+- updating the models to support the gemini 3.1 models.
+- improving the renitialization of the model